Create list4.ocd
Michael said I should make pull requests ;-), so here is one for testing the waters.

This is the list4 CD I described at the workshop in 2014, and have since made sporadic uses of in FMPs of other CDs (for example for the nassoc predicate symbol). It is deliberately exactly the file from that year, so that we may exercise (or not) bumping the version number and such; my point here is to create an example of the process, for future contributors to have a look at.